fix(deferred): adopt the deferred span cache list when span_count != 1
In the current version, rpmalloc will adopt deferred span cache list if
and only if the span_count == 1. This may cause memory leak in some
situations where the allocator thread allocate large blocks only and the
block is freed in another thread. So we try to adopt the deferred list
even when span_count != 1.

Besides, we also adopt deferred list when allocating huge blocks for the
same reason.