Add include for the private config
Normally this header is not available to code consuming the library,
however the test here is consuming it via add_subdirectory, which means
that it is also in the include path.

The test here requires the RPMALLOC_GROW_OR_FAIL variable, however that
is not in the public config header file, as it is not used for anything
in it.