Cross thread allocation/deallocation bug test + potential fixes
An unsynchronised access of span->list_size in
_rpmalloc_deallocate_direct_small_or_medium can result in a span being
recycled and put back into use before the store of
span->free_list_deferred occurs in the call to
_rpmalloc_deallocate_defer_small_or_medium that initially "freed" this
span. The eventual store of span->free_list_deferred corrupts the span
resulting in crashes.

This patch adds an artificial delay between the increment of
span->list_size and the store of span->free_list_deferred to make
reproduction of the issue more likely and thus faster with the included
test case.

Two potential fixes are also included.